# apps/api/src/schemas/user.py
"""Pydantic schemas for user operations"""
from pydantic import BaseModel, EmailStr, Field, field_validator
from datetime import datetime
from typing import Optional


class UserRegister(BaseModel):
    """Schema for user registration"""

    email: EmailStr
    password: str = Field(..., min_length=8)

    @field_validator('password')
    @classmethod
    def validate_password(cls, v: str) -> str:
        """Validate password strength"""
        if len(v) < 8:
            raise ValueError('Password must be at least 8 characters long')
        return v


class UserLogin(BaseModel):
    """Schema for user login"""

    email: EmailStr
    password: str


class UserResponse(BaseModel):
    """Schema for user response (excludes sensitive data)"""

    id: str
    email: str
    verified: bool = False
    created_at: datetime

    model_config = {
        "from_attributes": True,
    }


class UserVerify(BaseModel):
    """Schema for email verification"""

    token: str